Fix crash on dedicated server
